Tourism Website
Its a website for tourism development based on PHP and Wordpress
content. Its developed for the tourist who wanna free roam around the
Kathmandu valley. Tourist can search for hotels and their location. They
can see gallery of historic places. This project can only be run on
localhost and also the database file is attached in the last.